Added no-header feature in binary upload

+++ command/ota/massupload.go
@@ -21,7 +21,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 
@@ -37,11 +36,12 @@ const (
 // MassUploadParams contains the parameters needed to
 // perform a Mass OTA upload.
 type MassUploadParams struct {
-	DeviceIDs []string
-	Tags      map[string]string
-	File      string
-	Deferred  bool
-	FQBN      string
+	DeviceIDs        []string
+	Tags             map[string]string
+	File             string
+	Deferred         bool
+	DoNotApplyHeader bool
+	FQBN             string
 }
 
 // Result of an ota upload on a device.
@@ -59,17 +59,27 @@ func MassUpload(ctx context.Context, params *MassUploadParams, cred *config.Cred
 		return nil, errors.New("cannot use both DeviceIDs and Tags. only one of them should be not nil")
 	}
 
-	// Generate .ota file
-	otaDir, err := ioutil.TempDir("", "")
+	_, err := os.Stat(params.File)
 	if err != nil {
-		return nil, fmt.Errorf("%s: %w", "cannot create temporary folder", err)
+		return nil, fmt.Errorf("file %s does not exists: %w", params.File, err)
 	}
-	otaFile := filepath.Join(otaDir, "temp.ota")
-	defer os.RemoveAll(otaDir)
 
-	err = Generate(params.File, otaFile, params.FQBN)
-	if err != nil {
-		return nil, fmt.Errorf("%s: %w", "cannot generate .ota file", err)
+	// Generate .ota file
+	var otaFile string
+	if params.DoNotApplyHeader {
+		otaFile = params.File
+	} else {
+		otaDir, err := os.MkdirTemp("", "")
+		if err != nil {
+			return nil, fmt.Errorf("%s: %w", "cannot create temporary folder", err)
+		}
+		otaFile := filepath.Join(otaDir, "temp.ota")
+		defer os.RemoveAll(otaDir)
+
+		err = Generate(params.File, otaFile, params.FQBN)
+		if err != nil {
+			return nil, fmt.Errorf("%s: %w", "cannot generate .ota file", err)
+		}
 	}
 
 	iotClient, err := iot.NewClient(cred)
