+++ command/ota/upload.go
@@ -20,7 +20,6 @@ package ota
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 
@@ -38,14 +37,20 @@ const (
 // UploadParams contains the parameters needed to
 // perform an OTA upload.
 type UploadParams struct {
-	DeviceID string
-	File     string
-	Deferred bool
+	DeviceID         string
+	File             string
+	Deferred         bool
+	DoNotApplyHeader bool
 }
 
 // Upload command is used to upload a firmware OTA,
 // on a device of Arduino IoT Cloud.
 func Upload(ctx context.Context, params *UploadParams, cred *config.Credentials) error {
+	_, err := os.Stat(params.File)
+	if err != nil {
+		return fmt.Errorf("file %s does not exists: %w", params.File, err)
+	}
+
 	iotClient, err := iot.NewClient(cred)
 	if err != nil {
 		return err
@@ -56,16 +61,21 @@ func Upload(ctx context.Context, params *UploadParams, cred *config.Credentials)
 		return err
 	}
 
-	otaDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		return fmt.Errorf("%s: %w", "cannot create temporary folder", err)
-	}
-	otaFile := filepath.Join(otaDir, "temp.ota")
-	defer os.RemoveAll(otaDir)
+	var otaFile string
+	if params.DoNotApplyHeader {
+		otaFile = params.File
+	} else {
+		otaDir, err := os.MkdirTemp("", "")
+		if err != nil {
+			return fmt.Errorf("%s: %w", "cannot create temporary folder", err)
+		}
+		otaFile = filepath.Join(otaDir, "temp.ota")
+		defer os.RemoveAll(otaDir)
 
-	err = Generate(params.File, otaFile, dev.Fqbn)
-	if err != nil {
-		return fmt.Errorf("%s: %w", "cannot generate .ota file", err)
+		err = Generate(params.File, otaFile, dev.Fqbn)
+		if err != nil {
+			return fmt.Errorf("%s: %w", "cannot generate .ota file", err)
+		}
 	}
 
 	file, err := os.Open(otaFile)
